update_actions/news_requirements.txt
pandas==2.2.2
psycopg2-binary==2.9.9
requests==2.32.3
SQLAlchemy==2.0.31